Add function to use own style
A user can provide their own formatting style for bibliography entries with a script containing a new Style, which is based upon one of pybtex default classes.

Using PYBTEX_FORMAT_STYLE = 'myscriptname.MyCustomStyle' in the pelicanconf.py file then allows pelican-pybtex to use this style.

The function first checks if a default styles is used, then tries to import the custom style and falls back to plain if everything fails.